requirements.txt
# Core dependencies
SPARQLWrapper>=1.8.5
pandas>=1.3.0
requests>=2.25.0 
python-dotenv>=1.0.0

# LLM Provider packages (at least one is required)
# Comment out any provider you don't need
ollama>=0.4.0
openai>=1.7.0
anthropic>=0.5.0

# Other dependencies
tqdm>=4.65.0
logging>=0.4.9
typing-extensions>=4.7.1 